backend/deps/requirements.txt
dependency-injector
boto3
django
djangorestframework
celery
django-cors-headers
psycopg2
celery
redis
pandas